Remove (uint64_t) from encrypt.c

+++ src/Titanium_CCA_hi/encrypt.c
@@ -61,7 +61,7 @@ int crypto_encrypt_keypair(unsigned char *pk, unsigned char *sk, const unsigned 
 		/* a_i <-- a_i mp_{d+k} s */
 		for (j = 0; j < N + D + K1 + 1; j++)
 		{
-			a[i][j] = barrett_4q2((uint64_t)a[i][j] * (uint64_t)s[j]);
+			a[i][j] = barrett_4q2(a[i][j] * s[j]);
 		}
 		
 		INTT_NDK_DK(a[i]);
@@ -156,7 +156,7 @@ int crypto_encrypt(unsigned char *c, unsigned long long *clen, const unsigned ch
 	{
 		for (j = 0; j < N + K1 + 1; j++)
 		{
-			c1[j] = barrett_4q2((uint64_t)c1[j] + (uint64_t)r[i][j] * (uint64_t)a[i][j]);
+			c1[j] = barrett_4q2(c1[j] + r[i][j] * a[i][j]);
 		}
 	}
 	
@@ -166,7 +166,7 @@ int crypto_encrypt(unsigned char *c, unsigned long long *clen, const unsigned ch
 	{
 		for (j = 0; j < D + K1 + 1; j++)
 		{
-			c2[j] = barrett_4q2((uint64_t)c2[j] + (uint64_t)r2[i][j] * (uint64_t)b[i][j]);
+			c2[j] = barrett_4q2(c2[j] + r2[i][j] * b[i][j]);
 		}
 	}
 	
@@ -218,7 +218,7 @@ int crypto_encrypt_open(unsigned char *m, unsigned long long *mlen, const unsign
 	
 	for (i = 0; i < N + D + K1 + 1; i++)
 	{
-		c1[i] = barrett_4q2((uint64_t)c1[i] * (uint64_t)s[i]);
+		c1[i] = barrett_4q2(c1[i] * s[i]);
 	}
 	
 	INTT_NDK_D(c1);
