+++ src/Titanium_CCA_lite/encrypt.c
@@ -62,7 +62,7 @@ int crypto_encrypt_keypair(unsigned char *pk, unsigned char *sk, const unsigned 
 		/* a_i <-- a_i mp_{d+k} s */
 		for (j = 0; j < DIM_3; j++)
 		{
-			a[i][j] = barrett_4q2((uint64_t)a[i][j] * (uint64_t)s[j]);
+			a[i][j] = barrett_4q2(a[i][j] * s[j]);
 		}
 		
 		INTT_NDK_DK(a[i]);
@@ -159,13 +159,13 @@ int crypto_encrypt(unsigned char *c, unsigned long long *clen, const unsigned ch
 	/* c1 <-- NTT(sum(r_i * a_i)) */
 	for (j = 0; j < DIM_2; j++)
 	{
-		c1[j] = barrett_64q2((uint64_t)r[0][j] * (uint64_t)a[0][j] + (uint64_t)r[1][j] * (uint64_t)a[1][j] + (uint64_t)r[2][j] * (uint64_t)a[2][j] + (uint64_t)r[3][j] * (uint64_t)a[3][j] + (uint64_t)r[4][j] * (uint64_t)a[4][j] + (uint64_t)r[5][j] * (uint64_t)a[5][j] + (uint64_t)r[6][j] * (uint64_t)a[6][j] + (uint64_t)r[7][j] * (uint64_t)a[7][j] + (uint64_t)r[8][j] * (uint64_t)a[8][j]);
+		c1[j] = barrett_64q2(r[0][j] * a[0][j] + r[1][j] * a[1][j] + r[2][j] * a[2][j] + r[3][j] * a[3][j] + r[4][j] * a[4][j] + r[5][j] * a[5][j] + r[6][j] * a[6][j] + r[7][j] * a[7][j] + r[8][j] * a[8][j]);
 	}
 	
 	/* c2 <-- sum(r_i mp_d b_i) + mu */
 	for (j = 0; j < DIM_1; j++)
 	{
-		c2[j] = barrett_64q2((uint64_t)r2[0][j] * (uint64_t)b[0][j] + (uint64_t)r2[1][j] * (uint64_t)b[1][j] + (uint64_t)r2[2][j] * (uint64_t)b[2][j] + (uint64_t)r2[3][j] * (uint64_t)b[3][j] + (uint64_t)r2[4][j] * (uint64_t)b[4][j] + (uint64_t)r2[5][j] * (uint64_t)b[5][j] + (uint64_t)r2[6][j] * (uint64_t)b[6][j] + (uint64_t)r2[7][j] * (uint64_t)b[7][j] + (uint64_t)r2[8][j] * (uint64_t)b[8][j]);
+		c2[j] = barrett_64q2(r2[0][j] * b[0][j] + r2[1][j] * b[1][j] + r2[2][j] * b[2][j] + r2[3][j] * b[3][j] + r2[4][j] * b[4][j] + r2[5][j] * b[5][j] + r2[6][j] * b[6][j] + r2[7][j] * b[7][j] + r2[8][j] * b[8][j]);
 	}
 	
 	INTT_DK_D(c2);
@@ -216,7 +216,7 @@ int crypto_encrypt_open(unsigned char *m, unsigned long long *mlen, const unsign
 	
 	for (i = 0; i < DIM_3; i++)
 	{
-		c1[i] = barrett_4q2((uint64_t)c1[i] * (uint64_t)s[i]);
+		c1[i] = barrett_4q2(c1[i] * s[i]);
 	}
 	
 	INTT_NDK_D(c1);
